Revert "Stores from database"
