+++ backend/src/app.js
@@ -1,12 +1,11 @@
 import express from "express";
 import authRoutes from "./routes/authRoutes.js";
-import filtersRoutes from "./routes/filtersRoutes.js";
-import storesRoutes from "./routes/storesRoutes.js";
 import savedRoutes from "./routes/savedRoutes.js";
 import dotenv from "dotenv";
-import cors from "cors";
 
 dotenv.config();
+
+import cors from "cors";
 
 const app = express();
 
@@ -15,8 +14,6 @@ app.use(express.json());
 
 // routes
 app.use("/auth", authRoutes);
-app.use("/filters", filtersRoutes);
-app.use("/stores", storesRoutes);
 app.use("/route", savedRoutes)
 
 export default app;
